# apps/be/open-api-specs.json
{
  "openapi": "3.0.0",
  "paths": {
    "/funcionarios": {
      "post": {
        "operationId": "FuncionariosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFuncionarioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Funcionarios"
        ]
      },
      "get": {
        "operationId": "FuncionariosController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FuncionarioDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Funcionarios"
        ]
      }
    },
    "/funcionarios/{id}": {
      "get": {
        "operationId": "FuncionariosController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFuncionarioResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Funcionarios"
        ]
      },
      "patch": {
        "operationId": "FuncionariosController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFuncionarioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Funcionarios"
        ]
      },
      "delete": {
        "operationId": "FuncionariosController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Funcionarios"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    }
  },
  "info": {
    "title": "Dougust API",
    "description": "Auto-generated OpenAPI specification",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateFuncionarioDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "social": {
            "type": "string",
            "minLength": 0,
            "maxLength": 100
          },
          "cpf": {
            "type": "string",
            "minLength": 1,
            "maxLength": 15
          },
          "nascimento": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "minLength": 0,
            "maxLength": 20
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "rg": {
            "type": "string",
            "minLength": 0,
            "maxLength": 11
          }
        },
        "required": [
          "cpf",
          "email"
        ]
      },
      "FuncionarioDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "social": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "nascimento": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "rg": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nome",
          "social",
          "cpf",
          "nascimento",
          "phone",
          "email",
          "rg",
          "createdAt",
          "updatedAt"
        ]
      },
      "GetFuncionarioResponseDto": {
        "type": "object",
        "properties": {
          "funcionario": {
            "$ref": "#/components/schemas/FuncionarioDto"
          }
        },
        "required": [
          "funcionario"
        ]
      },
      "UpdateFuncionarioDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "social": {
            "type": "string",
            "minLength": 0,
            "maxLength": 100
          },
          "cpf": {
            "type": "string",
            "minLength": 1,
            "maxLength": 15
          },
          "nascimento": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "minLength": 0,
            "maxLength": 20
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "rg": {
            "type": "string",
            "minLength": 0,
            "maxLength": 11
          }
        }
      }
    }
  }
}